+++ LostPeterVulkan/LostPeterVulkan/src/EditorGrid.cpp
@@ -188,13 +188,13 @@ namespace LostPeterVulkan
     {
         StringVector* pDescriptorSetLayoutNames = this->pPipelineGraphics->poDescriptorSetLayoutNames;
         F_Assert(pDescriptorSetLayoutNames != nullptr && "EditorGrid::updateDescriptorSets_Graphics")
-        size_t count_ds = this->pPipelineGraphics->poDescriptorSets.size();
-        for (size_t i = 0; i < count_ds; i++)
+        uint32_t count_ds = (uint32_t)this->pPipelineGraphics->poDescriptorSets.size();
+        for (uint32_t i = 0; i < count_ds; i++)
         {
             VkWriteDescriptorSetVector descriptorWrites;
 
-            size_t count_names = pDescriptorSetLayoutNames->size();
-            for (size_t j = 0; j < count_names; j++)
+            uint32_t count_names = (uint32_t)pDescriptorSetLayoutNames->size();
+            for (uint32_t j = 0; j < count_names; j++)
             {
                 String& nameDescriptorSet = (*pDescriptorSetLayoutNames)[j];
                 if (nameDescriptorSet == Util_GetDescriptorSetTypeName(Vulkan_DescriptorSet_Pass)) //Pass
